+++ pkg/config/manager/BUILD
@@ -16,6 +16,7 @@ go_library(
         "//pkg/aspect:go_default_library",
         "//pkg/config:go_default_library",
         "//pkg/aspect/config:go_default_library",
+        "//pkg/cnfgNormalizer/go_default_library",
         "@com_github_ghodss_yaml//:go_default_library",
         "@com_github_gogo_protobuf//jsonpb:go_default_library",
         "@com_github_golang_glog//:go_default_library",
