+++ pkg/config/manager/manager.go
@@ -18,19 +18,17 @@ import (
 	"crypto/sha1"
 	"io/ioutil"
 	"sync"
-	"bytes"
 	"time"
 	"github.com/golang/glog"
 	"istio.io/mixer/pkg/config"
 	"istio.io/mixer/pkg/adapter"
-	aconfig "istio.io/mixer/pkg/aspect/config"
 	"istio.io/mixer/pkg/attribute"
 	"istio.io/mixer/pkg/config/descriptor"
+	"istio.io/mixer/pkg/cnfgNormalizer"
 	pb "istio.io/mixer/pkg/config/proto"
 
 
 	"istio.io/mixer/pkg/expr"
-	"fmt"
 )
 
 // Resolver resolves configuration to a list of combined configs.
@@ -38,7 +36,7 @@ type Resolver interface {
 	// Resolve resolves configuration to a list of combined configs.
 	Resolve(bag attribute.Bag, aspectSet config.AspectSet) ([]*pb.Combined, error)
 	// get JS string
-	GetJS() string
+	GetNormalizedConfig() *config.NormalizedConfig
 }
 
 // ChangeListener listens for config change notifications.
@@ -139,83 +137,10 @@ func (c *Manager) fetch() (*config.Runtime, descriptor.Finder, error) {
 	c.scSHA = scSHA
 	rt := config.NewRuntime(vd, c.eval)
 
-	rt.JSStr = getJS(vd, c)
+	rt.NormalizedConfig = cnfgNormalizer.Normalize(vd)
 	return rt, c.descriptorFinder, nil
 }
 
-func getJS(vd *config.Validated, c *Manager) string {
-	//vd.serviceConfig
-	/*
-	create methodStrs for each method (chk, report, quota)
-	for each rule r:
-	  create tmpMethodStrs for each method (chk, report, quota)
-          for each aspect a:
-	    generate JS k via Aspect Managers
-	    add k to appropriate tmpMethodStrs.
-	  for each non empty tmpMethodStrs {
-	    insert : if (rule) {
-	      tmpMethodStr
-	    }
-	  }
-	  Do recurrs for each nested rules.
-	*/
-	var reportMethodStr string
-	var injectionMethods string
-	for _, aspectRule := range vd.GetValidatedSC().GetRules() {
-		var tmpReportMethodStr string
-		for _, aspect := range aspectRule.GetAspects() {
-			if aspect.Kind == "metrics" {
-				invocation, wrapperMtdToIngest := GetJSInvocationForMetricAspect(aspect.Params.(*aconfig.MetricsParams), aspect.Adapter)
-				tmpReportMethodStr = tmpReportMethodStr + invocation
-				// HACK. skipping duplication incertion of mtd.
-				if len(injectionMethods) == 0 {
-					injectionMethods = injectionMethods + wrapperMtdToIngest
-				}
-			}
-		}
-		if len(tmpReportMethodStr) > 0 {
-			//TODO FIX FOR COMPLEXT OPTIONS
-			var ifStatementStr string
-			if len(aspectRule.Selector) > 0 {
-				expressionStr := aspectRule.Selector
-				ifStatementStr = getJSForExpression(expressionStr)
-
-			} else {
-				ifStatementStr = "true"
-			}
-			ruleIfBlocked := fmt.Sprintf(`
-			if (%s) {
-			  %s
-			}
-			`, ifStatementStr, tmpReportMethodStr)
-
-			reportMethodStr = reportMethodStr + ruleIfBlocked
-		}
-
-
-	}
-	allJSMethodFormat := `
-		function report(propertyBag) {
-		    %s
-		}
-		function check(propertyBag) {
-		    // TODO
-		}
-		function quota(propertyBag) {
-		    // TODO
-		}
-		`
-	userJSAllCode := fmt.Sprintf(allJSMethodFormat, reportMethodStr)
-	var userScript = userJSAllCode + "\n" + injectionMethods
-
-	fmt.Println(userScript)
-	err := ioutil.WriteFile("/tmp/generatedJSFromConfig.js", []byte(userScript), 0644)
-	if err != nil {
-		panic(err)
-	}
-
-	return userScript
-}
 // fetchAndNotify fetches a new config and notifies listeners if something has changed
 func (c *Manager) fetchAndNotify() error {
 	rt, df, err := c.fetch()
@@ -273,102 +198,4 @@ func (c *Manager) Start() {
 	if err != nil {
 		glog.Warning("Unable to process config: ", err)
 	}
-}
-
-///////////////////// THIS SHOULD BELONG TO METRIC ASPECT MANAGER /////////////////
-
-func GetJSInvocationForMetricAspect(metricsParams *aconfig.MetricsParams, adapterName string) (string, string) {
-	params := metricsParams
-	//var allMetricsStr bytes.Buffer
-	var metricsStr bytes.Buffer
-	for _, metric := range params.Metrics {
-		var labelStr bytes.Buffer
-		labelStr.WriteString(fmt.Sprintf("%s: %s", "value", getJSForExpression(metric.Value)))
-		labelLen := len(metric.Labels)
-		if labelLen != 0 {
-			labelStr.WriteString(",\n")
-		}
-		for key, value := range metric.Labels {
-			labelStr.WriteString(fmt.Sprintf(`				    %s: %s`, key, getJSForExpression(value)))
-			labelLen--
-			if labelLen != 0 {
-				labelStr.WriteString(",\n")
-			}
-		}
-		metricsStr.WriteString(fmt.Sprintf(`
-			      "%s": {
-			        %s
-			      },`, metric.DescriptorName, labelStr.String()))
-
-	}
-	metricsStrBuilt := metricsStr.String()
-	if metricsStrBuilt[len(metricsStrBuilt)-1] == ',' {
-		metricsStrBuilt = metricsStrBuilt[0 : len(metricsStrBuilt)-1]
-	}
-	callStr := fmt.Sprintf(`
-				%s({
-	  			  %s
-				})
-	`, GetJSWrapperMethodNameForMetricAspect(adapterName), metricsStrBuilt)
-	return callStr, GetJSWrapperMethodsToInjectForMetricAspect(adapterName, "metrics")
-}
-
-func GetJSWrapperMethodNameForMetricAspect(adapterName string) string {
-	return "RecordTo" + adapterName
-}
-
-func GetJSWrapperMethodsToInjectForMetricAspect(adapterName string, kindName string) string {
-
-	methodName := GetJSWrapperMethodNameForMetricAspect(adapterName)
-	var embeddedMethodsInUserScriptFmt = `
-                function %s(val) {
-                  CallBackFromUserScript_go("%s", "%s", val)
-                }
-`
-	return fmt.Sprintf(embeddedMethodsInUserScriptFmt, methodName, kindName, adapterName)
-}
-
-func getJSForExpression(expression string) string{
-	ex, err := expr.Parse(expression)
-	var out string
-	if err != nil {
-		glog.Warning("Unable to parse : %s. %v. Setting expression to false", expression, err)
-		out = "false"
-	} else {
-		condition, _ := EvalJSExpession(ex, expr.FuncMap(), "propertyBag.Get")
-		out = condition
-	}
-	return out
-}
-
-// Eval evaluates the expression given an attribute bag and a function map.
-func EvalJSExpession(e *expr.Expression, fMap map[string]expr.FuncBase, getPropMtdName string) (string, error) {
-	if e.Const != nil {
-		return e.Const.StrValue, nil
-	}
-	if e.Var != nil {
-		return fmt.Sprintf(getPropMtdName + "(\"%s\")[0]", e.Var.Name), nil
-	}
-
-	fn := fMap[e.Fn.Name]
-	if fn == nil {
-		return "", fmt.Errorf("unknown function: %s", e.Fn.Name)
-	}
-	// may panic
-	if e.Fn.Name == "EQ" {
-		leftStr, _ := EvalJSExpession(e.Fn.Args[0], fMap, getPropMtdName)
-		rightStr, _ := EvalJSExpession(e.Fn.Args[1], fMap, getPropMtdName)
-		return fmt.Sprintf("%s == %s", leftStr, rightStr), nil
-	}
-	if e.Fn.Name == "OR" {
-		//(age < 18) ? "Too young":"Old enough"
-		allArgs := e.Fn.Args
-		if len(allArgs) > 0 {
-			chkIfExists := fmt.Sprintf(getPropMtdName+"(\"%s\")[1]", allArgs[0].Var.Name)
-			leftexp, _ := EvalJSExpession(e.Fn.Args[0], fMap, getPropMtdName)
-			rightexp, _ := EvalJSExpession(e.Fn.Args[1], fMap, getPropMtdName)
-			return fmt.Sprintf("%s ? %s : %s", chkIfExists, leftexp, rightexp), nil
-		}
-	}
-	return "", nil
 }
